+++ src/resource/config_resource.cpp
@@ -5,6 +5,7 @@
 
 #include "config.h"
 #include "core/filesystem/file.h"
+#include "core/filesystem/filesystem.h"
 #include "core/json/json_object.inl"
 #include "core/json/sjson.h"
 #include "core/memory/allocator.h"
@@ -57,7 +58,17 @@ namespace config_resource_internal
 		DATA_COMPILER_ASSERT_RESOURCE_EXISTS("lua", boot_script.c_str(), opts);
 		DATA_COMPILER_ASSERT_RESOURCE_EXISTS("package", boot_package.c_str(), opts);
 
-		opts.write(buf);
+		if (opts._bundle) {
+			TempAllocator256 ta;
+			DynamicString dest(ta);
+			destination_path(dest, opts._resource_id);
+
+			File *config = opts._data_filesystem.open(dest.c_str(), FileOpenMode::READ);
+			file::copy(opts._file, *config, config->size());
+			opts._data_filesystem.close(*config);
+		} else {
+			opts.write(buf);
+		}
 
 		return 0;
 	}